+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Render `hello` on presence graph when unauthenticated.
+
+### Changed
+
+- Remove `beta` tag from parking button.
+
 ## [2024.5.2] - 2024-05-27
 
 ### Fixed